requirements.txt
flask
gradio
easyocr
pillow
requests
